+++ lsof-rs/crates/lsof-backend-windows/src/etw.rs
@@ -222,6 +222,7 @@ pub fn to_open_file(sock: &EtwSocket) -> OpenFile {
         format!("(endpoint 0x{:x})", sock.endpoint)
     };
     OpenFile {
+        lock: None,
         fd: FdType::Unknown,
         access: AccessMode::ReadWrite,
         file_type: sock.file_type(),